[PATCH] DTS: Speed up boot by skipping BogoMIPS calc
For a given clk frequency, the interAptiv in Pistachio will always
have the same performance, so there is no need to recalculate
BogoMIPS for every boot.

Instead pass in the result of the calculation.  This is based on a
CPU frequency of 546MHz.

(Note: This change is adopted from commit upstream ImgWorks 97e611343f82
"DTS: Speed up boot by skipping BogoMIPS calc")

This commit is just for completion, since we pass boot args via uboot.

Signed-off-by: Rahul Bedarkar <rahul.bedarkar@imgtec.com>