pistachio: image format re-write
This is an attempt to make the images follow the openwrt standard
rather than using the existing format that we have hacked into the
openwrt build system.

As part of this have moved to the Device build infrastructure so
that we can make use of it. Have also moved to fit images rather
than trying to support the legacy uImage format.

The old system had a single ubi volume containing a ubifs which in
turn contained the kernel (in uImage format) and the dtb in the
file system itself:

+----------------------------------------------------------+
|        rootfs(ubifs containing uImage and dtb)           |
+----------------------------------------------------------+
|                           UBI                            |
+----------------------------------------------------------+

The new system instead keeps the kernel in its own volume and also
now makes use of the overlay implementation already in openwrt which
uses a readonly squashfs and overlays it with a writable filesystem
at first boot.

+----------------------------------------------------------+
| kernel(fitImage) | rootfs(squashfs) |    rootfs_data     |
+----------------------------------------------------------+
|                           UBI                            |
+----------------------------------------------------------+

Note that this requires a new bootloader that expects this
layout and supports fit images.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>