image.mk: ubinize: remove hardcoded EOF option
This option doesn't seem to be part of the standard meaning if something else
(e.g. uboot) touches the nand the boot starts failing. Therefore have removed
it as a default and instead provided the ability to pass through extra args
as needed on a per device basis.

Note I don't know what devices need it so can't add it back in.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>